moved classes from challengehandlers to certbasedauth
