+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/certbasedauth/AbstractUsbSmartcardCertBasedAuthManager.java
@@ -20,18 +20,10 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.webview.challengehandlers;
-
-import lombok.Getter;
+package com.microsoft.identity.common.internal.ui.webview.certbasedauth;
 
 /**
- * An abstract smartcard CertBasedAuth manager specifically for NFC.
+ * An abstract smartcard CertBasedAuth manager specifically for USB.
  */
-public abstract class AbstractNfcSmartcardCertBasedAuthManager extends AbstractSmartcardCertBasedAuthManager<IConnectionCallback> {
-
-    /* boolean based on the most recently connected smartcard device
-    being the same as the previously connected smartcard device.
-    true if most recently connected smartcard device is the same as the previous.
-    false otherwise (including if no previous smartcard device exists). */
-    @Getter protected boolean isDeviceChanged;
+public abstract class AbstractUsbSmartcardCertBasedAuthManager extends AbstractSmartcardCertBasedAuthManager<IUsbConnectionCallback> {
 }
