+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/certbasedauth/ISmartcardSession.java
@@ -20,10 +20,9 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.webview.challengehandlers;
+package com.microsoft.identity.common.internal.ui.webview.certbasedauth;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
 import java.security.PrivateKey;
 import java.util.List;
