+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/certbasedauth/YubiKitSmartcardSession.java
@@ -20,15 +20,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.webview.challengehandlers;
+package com.microsoft.identity.common.internal.ui.webview.certbasedauth;
 
 import static com.yubico.yubikit.piv.Slot.AUTHENTICATION;
 import static com.yubico.yubikit.piv.Slot.CARD_AUTH;
 import static com.yubico.yubikit.piv.Slot.KEY_MANAGEMENT;
 import static com.yubico.yubikit.piv.Slot.SIGNATURE;
 
-
-import android.nfc.Tag;
 
 import androidx.annotation.NonNull;
